# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for LKE (Lembar Kerja Evaluasi) Application - Indonesian Government Evaluation System",
        "title": "LKE Application API",
        "termsOfService": "http://swagger.io/terms",
        "contact": {
            "name": "API Support",
            "email": "support@lke-app.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.6.14"
    },
    "paths": {
        "/v1/article": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new article",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Articles"
                ],
                "summary": "Create article",
                "parameters": [
                    {
                        "description": "Article data",
                        "name": "createForm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.CreateArticleForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "406": {
                        "description": "Validation failed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/article/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a specific article by ID for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Articles"
                ],
                "summary": "Get article by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Article ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Article not found or invalid parameter",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update an existing article by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Articles"
                ],
                "summary": "Update article",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Article ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated article data",
                        "name": "updateForm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.CreateArticleForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article updated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Invalid parameter",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "406": {
                        "description": "Validation failed or could not update",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete an article by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Articles"
                ],
                "summary": "Delete article",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Article ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article deleted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Invalid parameter",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "406": {
                        "description": "Could not delete",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/articles": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all articles for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Articles"
                ],
                "summary": "Get all articles",
                "responses": {
                    "200": {
                        "description": "Articles list",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "406": {
                        "description": "Could not get articles",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/error-logs": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieve recent application error logs for debugging",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Application"
                ],
                "summary": "Get recent error logs",
                "responses": {
                    "200": {
                        "description": "Recent error logs",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to get error logs",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/lke-evaluasi": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new LKE Evaluasi record or update existing one. Supports file upload for PUT requests.",
                "consumes": [
                    "application/json",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Used in E-Office",
                    "LKE Evaluasi"
                ],
                "summary": "Create or Update LKE Evaluasi",
                "parameters": [
                    {
                        "description": "LKE Evaluasi data",
                        "name": "createForm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.CreateLkeEvaluasiForm"
                        }
                    },
                    {
                        "type": "file",
                        "description": "File to upload (for PUT requests only)",
                        "name": "file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "LKE Evaluasi created or updated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "File required for PUT or invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "406": {
                        "description": "Validation failed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to upload file or update rekap values",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new LKE Evaluasi record or update existing one. Supports file upload for PUT requests.",
                "consumes": [
                    "application/json",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Used in E-Office",
                    "LKE Evaluasi"
                ],
                "summary": "Create or Update LKE Evaluasi",
                "parameters": [
                    {
                        "description": "LKE Evaluasi data",
                        "name": "createForm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.CreateLkeEvaluasiForm"
                        }
                    },
                    {
                        "type": "file",
                        "description": "File to upload (for PUT requests only)",
                        "name": "file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "LKE Evaluasi created or updated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "File required for PUT or invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "406": {
                        "description": "Validation failed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to upload file or update rekap values",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/lke-evaluasi/signed-url/{lke_rekap_id}/{kode_evaluasi}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Generate a signed URL to access a file associated with an LKE Evaluasi record",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Used in E-Office",
                    "LKE Evaluasi"
                ],
                "summary": "Get signed URL for LKE Evaluasi file",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "LKE Rekap ID",
                        "name": "lke_rekap_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Kode Evaluasi",
                        "name": "kode_evaluasi",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Signed URL for file access",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Missing required parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to generate signed URL",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/lke-evaluasi/sync-evaluasi": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Synchronize jawaban data into evaluasi records for the specified LKE Rekap ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Used in E-Office",
                    "LKE Evaluasi"
                ],
                "summary": "Sync LKE Evaluasi",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "LKE Rekap ID",
                        "name": "lke_rekap_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully synced evaluasi from jawaban",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid lke_rekap_id parameter",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to sync evaluasi or update rekap values",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/v1/lke-evaluasi/{id}/{format}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a specific LKE Evaluasi record by ID for the authenticated user",
                "produces": [
                    "application/json",
                    "text/xml"
                ],
                "tags": [
                    "LKE Evaluasi"
                ],
                "summary": "Get LKE Evaluasi by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "LKE Evaluasi ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Response format: json or xml (defaults to json)",
                        "name": "format",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "LKE Evaluasi data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "LKE Evaluasi not found or invalid parameter",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/lke-evaluasis": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all LKE Evaluasi records for the authenticated user",
                "produces": [
                    "application/json",
                    "text/xml"
                ],
                "tags": [
                    "LKE Evaluasi"
                ],
                "summary": "Get all LKE Evaluasi records",
                "responses": {
                    "200": {
                        "description": "LKE Evaluasi records list",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "406": {
                        "description": "Could not get LKE Evaluasi records",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/lke-evaluasis/{format}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all LKE Evaluasi records for the authenticated user",
                "produces": [
                    "application/json",
                    "text/xml"
                ],
                "tags": [
                    "LKE Evaluasi"
                ],
                "summary": "Get all LKE Evaluasi records",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Response format: json or xml (defaults to json)",
                        "name": "format",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "LKE Evaluasi records list",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "406": {
                        "description": "Could not get LKE Evaluasi records",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/v1/user/logout": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Logout user and invalidate access token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "User logout",
                "responses": {
                    "200": {
                        "description": "Successfully logged out",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "forms.CreateArticleForm": {
            "type": "object",
            "required": [
                "content",
                "title"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 3
                },
                "title": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}